# perf(sorting): Sorting performance optimizations