--- a/.github/workflows/nodejs.yml
+++ b/.github/workflows/nodejs.yml
@@ -8,6 +8,7 @@ on:
     branches: [ master, '[0-9]+.[0-9]+.x' ]
   pull_request:
     branches: [ master, '[0-9]+.[0-9]+.x' ]
+  workflow_dispatch:
 
 jobs:
   run-tests: